chore(storage): add statistics test and catalog test
