+++ src/storage/secondary/block/char_block_builder.rs
@@ -2,7 +2,8 @@
 
 use risinglight_proto::rowset::BlockStatistics;
 
-use super::{BlockBuilder, StatisticsBuilder};
+use super::super::statistics::StatisticsBuilder;
+use super::BlockBuilder;
 use crate::array::Utf8Array;
 
 /// Encodes fixed-width char into a block.
