+++ tests/sql/catalog.slt
@@ -1,0 +1,7 @@
+statement ok
+create table t(v1 int, v2 int, v3 int)
+
+query IIIIII
+\dt
+----
+0 postgres 0 postgres 0 t
