+++ tests/sql/statistics.slt
@@ -1,0 +1,23 @@
+statement ok
+create table t(v1 int, v2 int, v3 int)
+
+statement ok
+insert into t values (1,10,100), (2,20,100), (3,10,100)
+
+query II
+\stat t v1
+----
+RowCount Int64(3)
+DistinctValue Int64(3)
+
+query II
+\stat t v2
+----
+RowCount Int64(3)
+DistinctValue Int64(2)
+
+query II
+\stat t v3
+----
+RowCount Int64(3)
+DistinctValue Int64(1)
