+++ tests/sqllogictest.rs
@@ -25,6 +25,7 @@ use tokio::runtime::Runtime;
 #[test_case("copy/csv.slt")]
 #[test_case("where.slt")]
 #[test_case("tpch/tpch.slt")]
+#[test_case("catalog.slt")]
 // #[test_case("select.slt")]
 // #[test_case("issue_347.slt")]
 fn test_mem(name: &str) {
@@ -53,6 +54,8 @@ fn test_mem(name: &str) {
 #[test_case("copy/csv.slt")]
 #[test_case("where.slt")]
 #[test_case("tpch/tpch.slt")]
+#[test_case("statistics.slt")]
+#[test_case("catalog.slt")]
 // #[test_case("select.slt")]
 // #[test_case("issue_347.slt")]
 fn test_disk(name: &str) {
